A simple Tic-Tac-Toe game coded in the span of a few days, the reasoning behind the game was to kill time.
To play launch TicTacToe.jar, the first player either 'X' or 'O' is
randomly selected and the player selects from the selection box the
location of where they would like to go for their turn. When there is a
winner, the winning location is highlighted. To play again simply close
the app and run it again, sorry a restart function is not added.

Future plans:
- add AI so you can play against a computer
- add a score counter
- properly center the X's and O's
- add restart function